📦 NEW: Improved the existing taskserver UI
